Moved to separate classs example error mappers
